refactor(frontend): 拆分 ToolDebugDialog 主组件职责 (#3089)
按照 Issue #3089 的建议，将主组件的过多职责拆分到独立模块：

1. 提取 renderFormField 函数（~140行）到 form-field-renderer 模块
   - 创建 createFormFieldRenderer 工厂函数
   - 导出 getTypeBadgeColor 辅助函数
   - 保持递归渲染逻辑完整性

2. 提取键盘事件处理（~45行）到 useToolShortcut Hook
   - 封装 Cmd/Ctrl+Enter 快捷键处理
   - 统一事件监听器生命周期管理

3. 主组件简化效果：
   - 从 ~635 行降至 ~490 行
   - 职责更加清晰：状态管理、表单逻辑、事件处理、UI渲染

遵循项目"务实开发"理念，渐进式优化而非预防性重构。
