+++ apps/frontend/src/components/form-field-renderer/index.ts
@@ -1,0 +1,11 @@
+/**
+ * 工具调试对话框组件模块
+ *
+ * @description 该模块包含工具调试对话框及其相关的子组件和辅助函数
+ */
+
+export {
+  createFormFieldRenderer,
+  getTypeBadgeColor,
+} from "./render-form-field";
+export type { FormFieldRenderer } from "./render-form-field";
